+++ src/core/repoIdRegistry.js
