+++ src/homepage/routeData.js
@@ -8,7 +8,7 @@
 
 /*::
 import type {Assets} from "../webutil/assets";
-import type {RepoIdRegistry} from "../explorer/repoIdRegistry";
+import type {RepoIdRegistry} from "../core/repoIdRegistry";
 
 type RouteDatum = {|
   +path: string,
